requirements.txt
Flask
Flask-SocketIO
fpdf2